// src/services/openapi/models/PostDto.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { CommentDto } from './CommentDto';
import type { PostCandidatureDto } from './PostCandidatureDto';
import type { PostRateDto } from './PostRateDto';
import type { UserDto } from './UserDto';

export type PostDto = {
    id: number;
    createdAt: string;
    updatedAt: string;
    slug: string;
    image: string;
    link: string;
    participation: number;
    title: string;
    content: string;
    availlablePositions: number;
    userId: number;
    user: UserDto;
    comments: Array<CommentDto>;
    rates: Array<PostRateDto>;
    candidatures: Array<PostCandidatureDto>;
};